requirements.txt
asgiref==3.3.4
Django==3.2.4
django-resized==0.3.11
djangorestframework==3.12.4
jmespath==0.10.0
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
